sweeper endpoint will send records from all peers
