+++ comms/discovery/rpcz/peer_sweeper_test.go
@@ -3,10 +3,8 @@ package rpcz
 import (
 	"errors"
 	"fmt"
-	"log"
 	"testing"
 
-	"comms.audius.co/discovery/config"
 	"comms.audius.co/discovery/db"
 	"comms.audius.co/discovery/schema"
 	"github.com/stretchr/testify/assert"
@@ -14,22 +12,23 @@ import (
 
 func TestPeerClient(t *testing.T) {
 	db.Conn.MustExec(`truncate rpc_error`)
+	var err error
 
-	discoveryConfig := config.Parse()
+	// discoveryConfig := config.Parse()
 
-	proc, err := NewProcessor(discoveryConfig)
-	if err != nil {
-		log.Fatal(err)
-	}
+	// proc, err := NewProcessor(discoveryConfig)
+	// if err != nil {
+	// 	log.Fatal(err)
+	// }
 
-	pc := NewPeerClient("https://example.com", proc)
+	// pc := NewPeerClient("https://example.com", proc)
 
 	op := &schema.RpcLog{
 		Sig:       "sig1",
 		Rpc:       []byte(`{"fun": true}`),
 		RelayedBy: "https://example.com",
 	}
-	err = pc.insertRpcError(op, errors.New("error1"))
+	err = insertRpcError(op, errors.New("error1"))
 	assert.NoError(t, err)
 
 	var errorCount int
@@ -40,7 +39,7 @@ func TestPeerClient(t *testing.T) {
 	assert.Equal(t, "error1", errorText)
 
 	// second attempt
-	err = pc.insertRpcError(op, errors.New("error2"))
+	err = insertRpcError(op, errors.New("error2"))
 	assert.NoError(t, err)
 
 	err = db.Conn.QueryRow(`select error_count, error_text from rpc_error where sig = 'sig1'`).Scan(&errorCount, &errorText)
@@ -55,40 +54,37 @@ func TestPeerClient(t *testing.T) {
 			Rpc:       []byte(`{"fun": true}`),
 			RelayedBy: "https://example.com",
 		}
-		err = pc.insertRpcError(op, errors.New("error2"))
+		err = insertRpcError(op, errors.New("error2"))
 		assert.NoError(t, err)
 	}
 
-	// add a third error but from an unrelated host
-	// this peer client should only consider (https://example.com)
+	// add a third error but from an different host
 	{
 		op := &schema.RpcLog{
 			Sig:       "sig3",
 			Rpc:       []byte(`{"fun": true}`),
 			RelayedBy: "https://google.com",
 		}
-		err = pc.insertRpcError(op, errors.New("error3"))
+		err = insertRpcError(op, errors.New("error3"))
 		assert.NoError(t, err)
 	}
 
 	{
-		var myFailed []*schema.RpcLog
-		err := pc.appendRpcErrorRows(&myFailed)
+		myFailed, err := getRpcErrorRows()
 		assert.NoError(t, err)
-		assert.Len(t, myFailed, 2)
+		assert.Len(t, myFailed, 3)
 	}
 
 	// increment retry above threshold
 	for i := 0; i < 100; i++ {
-		pc.insertRpcError(op, fmt.Errorf("error%d", i))
+		insertRpcError(op, fmt.Errorf("error%d", i))
 	}
 
 	// should ignore error that is over threshold
 	{
-		var myFailed []*schema.RpcLog
-		err := pc.appendRpcErrorRows(&myFailed)
+		myFailed, err := getRpcErrorRows()
 		assert.NoError(t, err)
-		assert.Len(t, myFailed, 1)
+		assert.Len(t, myFailed, 2)
 	}
 
 }
