+++ comms/discovery/server/server.go
@@ -746,8 +746,8 @@ func (ss *ChatServer) getRpcBulk(c echo.Context) error {
 		fmt.Println("failed to parse time", err, c.QueryParam("after"), c.QueryString())
 	}
 
-	query := `select * from rpc_log where relayed_by = $1 and relayed_at > $2 order by relayed_at asc limit 10000`
-	err := db.Conn.Select(&rpcs, query, ss.config.MyHost, after.Truncate(time.Microsecond))
+	query := `select * from rpc_log where applied_at > $1 order by applied_at asc limit 10000`
+	err := db.Conn.Select(&rpcs, query, after.Truncate(time.Microsecond))
 	if err != nil {
 		return err
 	}
